@@ -7,8 +7,10 @@
 		"test": "jest"
 	},
 	"dependencies": {
-		"react": "16.0.0-alpha.6",
-		"react-native": "0.44.1",
+		"metro-react-native-babel-preset": "^0.48.0",
+		"prop-types": "15.6.0",
+		"react": "16.5.0",
+		"react-native": "0.57",
 		"react-native-wkwebview-reborn": "file:../"
 	},
 	"devDependencies": {
