@@ -8,19 +8,22 @@ import {
 import WkWebView from 'react-native-wkwebview-reborn';
 
 export default class example extends Component {
-  componentWillMount() {
-    
-  }
   render() {
     return (
       <View style={{ flex: 1, marginTop: 20 }}>
-        <WkWebView style={{ backgroundColor: '#ff0000' }} 
+        <WkWebView style={{ backgroundColor: '#ff0000' }}
+          contentInsetAdjustmentBehavior="always"
           userAgent="MyFancyWebView"
           hideKeyboardAccessoryView={false}
-          ref="webview" 
-          sendCookies={true} 
-          source={{ uri: 'https://httpbin.org/get' }} />
-        <Text style={{ fontWeight: 'bold', padding: 10 }} onPress={() => this.refs.webview.reload()}>Reload</Text>
+          ref={(c) => this.webview = c}
+          sendCookies={true}
+          source={{ uri: 'https://example.org/' }}
+          onMessage={(e) => console.log(e.nativeEvent)}
+          onNavigationResponse={(e) => console.log(e.nativeEvent)}
+          injectedJavaScript="window.postMessage('Hello from WkWebView'); document.addEventListener('message', function(e) { alert(e.data); });"
+        />
+        <Text style={{ fontWeight: 'bold', padding: 10 }} onPress={() => this.webview.reload()}>Reload</Text>
+        <Text style={{ fontWeight: 'bold', padding: 10 }} onPress={() => this.webview.postMessage("Hello from React Native")}>Post Message</Text>
       </View>
     );
   }
