@@ -2,7 +2,7 @@
 
 #import <React/RCTView.h>
 
-@class RCTWKWebView;
+@class CRAWKWebView;
 
 /**
  * Special scheme used to pass messages to the injectedJavaScript
@@ -12,26 +12,32 @@
  */
 extern NSString *const RCTJSNavigationScheme;
 
-@protocol RCTWKWebViewDelegate <NSObject>
+@protocol CRAWKWebViewDelegate <NSObject>
 
-- (BOOL)webView:(RCTWKWebView *)webView
+- (BOOL)webView:(CRAWKWebView *)webView
 shouldStartLoadForRequest:(NSMutableDictionary<NSString *, id> *)request
    withCallback:(RCTDirectEventBlock)callback;
 
 @end
 
-@interface RCTWKWebView : RCTView
+@interface CRAWKWebView : RCTView
 
 - (instancetype)initWithProcessPool:(WKProcessPool *)processPool;
 
-@property (nonatomic, weak) id<RCTWKWebViewDelegate> delegate;
+@property (nonatomic, weak) id<CRAWKWebViewDelegate> delegate;
 
 @property (nonatomic, copy) NSDictionary *source;
 @property (nonatomic, assign) UIEdgeInsets contentInset;
 @property (nonatomic, assign) BOOL automaticallyAdjustContentInsets;
+@property (nonatomic, assign) BOOL messagingEnabled;
+@property (nonatomic, assign) BOOL allowsLinkPreview;
 @property (nonatomic, assign) BOOL openNewWindowInWebView;
+@property (nonatomic, assign) BOOL injectJavaScriptForMainFrameOnly;
+@property (nonatomic, assign) BOOL injectedJavaScriptForMainFrameOnly;
+@property (nonatomic, copy) NSString *injectJavaScript;
 @property (nonatomic, copy) NSString *injectedJavaScript;
 @property (nonatomic, assign) BOOL hideKeyboardAccessoryView;
+@property (nonatomic, assign) BOOL keyboardDisplayRequiresUserAction;
 
 
 - (void)goForward;
@@ -40,6 +46,7 @@ shouldStartLoadForRequest:(NSMutableDictionary<NSString *, id> *)request
 - (BOOL)canGoForward;
 - (void)reload;
 - (void)stopLoading;
+- (void)postMessage:(NSString *)message;
 - (void)evaluateJavaScript:(NSString *)javaScriptString completionHandler:(void (^)(id, NSError *error))completionHandler;
 
 @end
